requirements.txt
Django>=2.2.2,<2.3.0
djangorestframework>=3.9.4,<3.10.0
flake8>=3.7.7,<3.8.0
psycopg2>=2.8.2,<2.9.0
django-phonenumber-field>=3.0.1,<3.1.0
phonenumbers>=8.10.13,<8.11.0
stripe>=2.32.1,<2.33.0
Pillow>=6.1.0,<6.2.0
django-filter>=2.1.0,<2.2.0
